+++ AlegriaApp/urls.py
@@ -19,7 +19,7 @@ from rest_framework import permissions
 from drf_yasg.views import get_schema_view
 from drf_yasg import openapi
 
-from ApiArriendosAlegria.views import Login, Logout
+from ApiArriendosAlegria.views.auth_views import Login, Logout
 
 
 schema_view = get_schema_view(
@@ -34,7 +34,6 @@ schema_view = get_schema_view(
    public=True,
    permission_classes=[permissions.AllowAny],
 )
-
 
 
 urlpatterns = [
