+++ ApiArriendosAlegria/permission.py
@@ -1,5 +1,6 @@
 from rest_framework.permissions import BasePermission
 
+
 class IsAdminUser(BasePermission):
     """
     Allows access only to SuperUser users.
@@ -8,6 +9,7 @@ class IsAdminUser(BasePermission):
     def has_permission(self, request, view):
         return bool(request.user and request.user.is_superuser)
     
+
 class IsStaffUser(BasePermission):
     """
     Allows access only to Staff users.
