+++ ApiArriendosAlegria/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path, include
 
 urlpatterns = [
-    path('', include('ApiArriendosAlegria.routers'), name='trabajador_router'),
+    path('', include('ApiArriendosAlegria.routers'), name='api_router'),
 ]
