+++ ApiArriendosAlegria/views/api_views.py
@@ -1,19 +1,16 @@
 from datetime import datetime
 from dateutil.relativedelta import relativedelta
-from django.http import JsonResponse
-from rest_framework.decorators import action
+
 from django.utils import timezone
-from dateutil import tz
-from rest_framework.viewsets import GenericViewSet
-from django.contrib.sessions.models import Session
+
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import status, viewsets, generics
+
+from rest_framework import status, viewsets
+from rest_framework.decorators import action
 from rest_framework.response import Response
-from rest_framework.views import APIView
-from rest_framework.authtoken.models import Token
-from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.permissions import IsAuthenticated
-from ApiArriendosAlegria.fecha_scl import GetfechaScl
+
+from ApiArriendosAlegria.utils import GetfechaScl
 
 from ApiArriendosAlegria.models import (
     Usuario,
@@ -37,7 +34,7 @@ from ApiArriendosAlegria.models import (
     ValoresGlobales,
     CodigoPropiedad
 )
-from ApiArriendosAlegria.serializers import (
+from ApiArriendosAlegria.serializers.base_serializers import (
     SerializadorUsuario,
     SerializerArrendatarioArriendo,
     SerializerArriendoDepartamento,
@@ -62,145 +59,30 @@ from ApiArriendosAlegria.serializers import (
     SerializerActualizarValorArriendo,
     SerializerArriendoConDetalles,
     ListadoCodigoPropiedadSerializer,
-    ArriendMultaDashboardSerializer
 )
-# from django.db import transaction
+
 from ApiArriendosAlegria.permission import IsStaffUser
 from ApiArriendosAlegria.authentication_mixins import Authentication
 
 
-
-# --- General views: Login / Logout ---
-class Login(ObtainAuthToken):
-    """
-    Vista Login.
-
-    Usa autenticación de token propia de Django REST Framework (Authtoken).
-
-    Se admite sólo una sesión activa por usuario. Dicha sesión se destruye
-    si se repite la petición POST con la sesión ya iniciada.
-    """
-    def post(self, request, *args, **kwargs):
-
-        login_serializer = self.serializer_class(
-            data=request.data, context={'request': request})
-        
-        if not login_serializer.is_valid():
-            return Response({'error': 'Usuario o contraseña incorrecta'},
-                        status=status.HTTP_400_BAD_REQUEST)
-
-        user = login_serializer.validated_data['user']
-        if not user.is_active:
-            return Response({'error': 'No se puede ingresar con usuario inactivo'},
-                        status=status.HTTP_401_UNAUTHORIZED)
-        
-        token, created = Token.objects.get_or_create(user=user)
-        user_serializer = SerializadorUsuario(user)
-
-        if not created:
-            # Delete user token
-            token.delete()
-            # Delete all sessions for user
-            all_sessions = Session.objects.filter(
-                expire_date__gte=datetime.now())
-            
-            if all_sessions.exists():
-                for session in all_sessions:
-                    session_data = session.get_decoded()
-                    # auth_user_id is the primary key's user on the session
-                    if user.id == int(session_data.get('_auth_user_id')):
-                        session.delete()
-
-            return Response({'error': 'Su sessión quedo activa desde la última vez. Por favor ingrese nuevamente.'},
-                            status=status.HTTP_400_BAD_REQUEST
-                            )
-    
-        return Response({
-                    'token': token.key,
-                    'usuario': user_serializer.data,
-                    'message': 'Ingreso exitoso.'
-                }, status=status.HTTP_201_CREATED)
-
-        
-
-        
-
-class Logout(APIView):
-    """
-    Vista Logout.
-    
-    Logout mediante authtoken de Django REST Framework.
-    """
-    def post(self, request, *args, **kwargs):
-        try:
-            token_request = request.GET.get('token')
-            token = Token.objects.filter(key=token_request).first()
-
-            if not token:
-                return Response({'error': 'No hay usuario con ese token'},
-                            status=status.HTTP_400_BAD_REQUEST)
-            
-            user = token.user
-
-            token.delete()
-            # Delete all sessions for user
-            all_sessions = Session.objects.filter(
-                expire_date__gte=datetime.now())
-            if all_sessions.exists():
-                for session in all_sessions:
-                    session_data = session.get_decoded()
-                    # auth_user_id is the primary key's user on the session
-                    if user.id == int(session_data.get('_auth_user_id')):
-                        session.delete()
-            # Delete user token
-
-            token_message = 'Token eliminado.'
-            session_message = 'Sesión exitosamente cerrada.'
-
-            return Response({'token_message': token_message,
-                            'session_message': session_message},
-                            status=status.HTTP_200_OK)
-
-        except:
-            return Response({'error': 'No se ha encontrado el token ingresado'},
-                            status=status.HTTP_409_CONFLICT)     
-            
-            
-
 # -------------Api Bancos---------------
 class BancoViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    Vista "Banco".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated]
-
     serializer_class = SerializerBanco
     queryset = Banco.objects.all()
 
+
 # -------------Api Tipo Cuentas Bancarias---------------
 class TipoCuentaBancariaViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    Vista "Tipo Cuenta" (Bancaria).
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated]
-
     serializer_class = SerializerTipoCuenta
     queryset = TipoCuenta.objects.all()
 
 
 # -------------Api TypeWorkers---------------
 class TypeWorkerViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Tipo Trabajador".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerTipoTrabajado
@@ -209,11 +91,6 @@ class TypeWorkerViewSet(viewsets.ModelViewSet):
 
 # -------------Api Worker---------------
 class TrabajadorViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Trabajador".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerTrabajador
@@ -224,24 +101,15 @@ class TrabajadorViewSet(viewsets.ModelViewSet):
     
 # -------------Api Regiones--------------- 
 class RegionReadOnlyViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    Vista "Region".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated]
     queryset = Region.objects.all()
     serializer_class = SerializerRegion
     filter_backends = [DjangoFilterBackend]
 
+
 # -------------Api Communes---------------   
 class ComunaReadOnlyViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    Vista "Comuna".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     queryset = Comuna.objects.all()
@@ -252,11 +120,6 @@ class ComunaReadOnlyViewSet(viewsets.ReadOnlyModelViewSet):
 
 # --- API Usuario (nuevo) ---
 class UsuarioViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Usuario".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializadorUsuario
@@ -265,11 +128,6 @@ class UsuarioViewSet(viewsets.ModelViewSet):
 
 # ---------------------Segundo sprint-------------------
 class PropietarioViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Propietario".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerPropietario
@@ -289,11 +147,6 @@ class PropietarioViewSet(viewsets.ModelViewSet):
     
     
 class PersonalidadJuridicaViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Personalidad Jurídica".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerPersonalidadJuridica
@@ -301,11 +154,6 @@ class PersonalidadJuridicaViewSet(viewsets.ModelViewSet):
 
 
 class CuentaViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Cuenta".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerCuenta
@@ -314,11 +162,6 @@ class CuentaViewSet(viewsets.ModelViewSet):
     filterset_fields = ['cuenta','propietario_rut']
 
 class PropiedadViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Propiedad".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerPropiedad
@@ -336,22 +179,12 @@ class PropiedadViewSet(viewsets.ModelViewSet):
         return super().list(request, *args, **kwargs)
 
 class TipoPropiedadViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Tipo Propiedad".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerTipoPropiedad
     queryset = TipoPropiedad.objects.all()
 
 class ArriendatarioViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Arrendatario".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerArrendatario
@@ -371,24 +204,12 @@ class ArriendatarioViewSet(viewsets.ModelViewSet):
 
     
 class ArriendoViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Arriendo".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerArriendo
     queryset = Arriendo.objects.all()
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ['estado_arriendo','propiedad']
-    
-    # @action(detail=False methods=['get'])
-    # def dashboard_multas(self, request):
-    # # ArriendMultaDashboardSerializer
-    #     arriendo_atrasados = Arriendo.objects.filter()
-    #     serializer_class = ArriendMultaDashboardSerializer()
-    #     return Response
 
     def get_serializer_class(self):
         if self.action == 'list':
@@ -426,22 +247,9 @@ class ArriendoViewSet(viewsets.ModelViewSet):
         headers = self.get_success_headers(serializer.data)
         
         return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
-
-    # def destroy(self, request, *args, **kwargs):
-    #     instance = self.get_object()
-    #     if instance.externo:
-    #         instance.externo.delete()
-            
-    #     self.perform_destroy(instance)
-    #     return Response(status=status.HTTP_204_NO_CONTENT)
     
     
 class ArriendoDepartamentoViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Arriendo departamento".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerArriendoDepartamento
@@ -449,7 +257,7 @@ class ArriendoDepartamentoViewSet(viewsets.ModelViewSet):
     
 
 
-class DashboardViewSet(GenericViewSet):
+class DashboardViewSet(viewsets.GenericViewSet):
     queryset = DetalleArriendo.objects.all() 
     serializer_class = SerializerDetalleArriendo
     
@@ -519,33 +327,12 @@ class DashboardViewSet(GenericViewSet):
 
 
 class DetalleArriendoViewSet(viewsets.ModelViewSet):
-    """
-    Vista "Detalle Arriendo".
-
-    Métodos disponibles: list, create, retrieve, update, destroy.
-    """
     authentication_classes = [Authentication]
     permission_classes = [IsAuthenticated, IsStaffUser]
     serializer_class = SerializerDetalleArriendo
     queryset = DetalleArriendo.objects.all()
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ['arriendo']
-
-    """
-    @action(detail=True, methods=['post'])
-    def set_password(self, request, pk=None):
-        user = self.get_object()
-        serializer = PasswordSerializer(data=request.data)
-        if serializer.is_valid():
-            user.set_password(serializer.validated_data['password'])
-            user.save()
-            return Response({'status': 'password set'})
-        else:
-            return Response(serializer.errors,
-                            status=status.HTTP_400_BAD_REQUEST)
-                            
-                            
-    """
   
   
     @action(detail=True, methods=['post'])
